+++ inc/PROsyst.h
@@ -33,7 +33,7 @@ namespace PROfit {
             PROsyst excluding(const std::vector<std::string> &systs);
 
             /* Function: given the systematic name, return corresponding fractional covariance matrix */
-            Eigen::MatrixXd GrabMatrix(const std::string& sys) const;
+            Eigen::MatrixXf GrabMatrix(const std::string& sys) const;
 
             /* Function: given the systematic name, return corresponding Spline */
             Spline GrabSpline(const std::string& sys) const;
@@ -50,13 +50,13 @@ namespace PROfit {
             //----- Spline and Covariance matrix related ---
             //----- Spline and Covariance matrix related ---
 
-            Eigen::MatrixXd SumMatrices() const;
-            Eigen::MatrixXd SumMatrices(const std::vector<std::string>& sysnames) const;
+            Eigen::MatrixXf SumMatrices() const;
+            Eigen::MatrixXf SumMatrices(const std::vector<std::string>& sysnames) const;
 
             /* Function: given a SystStruct with cv and variation spectra, build full covariance matrix for the systematics, and return it
              * Note: it assumes the SystStruct is filled 
              */
-            static Eigen::MatrixXd GenerateFullCovarMatrix(const SystStruct& sys_obj);
+            static Eigen::MatrixXf GenerateFullCovarMatrix(const SystStruct& sys_obj);
 
             /* Function: Given a SystStruct, generate fractinal covariance matrix, and correlation matrix, and add matrices to covmat_map and corrtmat_map
              * Note: this function is lazy. It wouldn't do anything if it found covariance matrix with the same name already in the map.
@@ -66,29 +66,29 @@ namespace PROfit {
             /* Function: given a syst struct with cv and variation spectra, build fractional covariance matrix for the systematics, as well as correlation matrix 
              * Return: {fractional covariance matrix, correlation covariance matrix}
              */
-            static std::pair<Eigen::MatrixXd, Eigen::MatrixXd> GenerateCovarMatrices(const SystStruct& sys_obj);
+            static std::pair<Eigen::MatrixXf, Eigen::MatrixXf> GenerateCovarMatrices(const SystStruct& sys_obj);
 
             /* Function: given a SystStruct with cv and variation spectra, build fractional covariance matrix for the systematics, and return it
              * Note: it assumes the SystStruct is filled 
              */
-            static Eigen::MatrixXd GenerateFracCovarMatrix(const SystStruct& sys_obj);
+            static Eigen::MatrixXf GenerateFracCovarMatrix(const SystStruct& sys_obj);
 
             /* Given fractional covariance matrix, calculate the correlation matrix */
-            static Eigen::MatrixXd GenerateCorrMatrix(const Eigen::MatrixXd& frac_matrix);
+            static Eigen::MatrixXf GenerateCorrMatrix(const Eigen::MatrixXf& frac_matrix);
 
             /* Function: check if matrix has nan, or infinite value */
-            static bool isFiniteMatrix(const Eigen::MatrixXd& in_matrix);
+            static bool isFiniteMatrix(const Eigen::MatrixXf& in_matrix);
 
             /* Function: if matrix has nan/inf values, change to 0. 
              * Note: this modifies the matrix !! 
              */
-            static void toFiniteMatrix(Eigen::MatrixXd& in_matrix);
+            static void toFiniteMatrix(Eigen::MatrixXf& in_matrix);
 
             /* Function: check if given matrix is positive semi-definite with tolerance. UST THIS ONE!!*/
-            static bool isPositiveSemiDefinite_WithTolerance(const Eigen::MatrixXd& in_matrix, double tolerance=1.0e-16);
+            static bool isPositiveSemiDefinite_WithTolerance(const Eigen::MatrixXf& in_matrix, float tolerance=1.0e-16);
 
             /* Function: check if given matrix is positive semi-definite, no tolerance at all (besides precision error from Eigen) */
-            static bool isPositiveSemiDefinite(const Eigen::MatrixXd& in_matrix);
+            static bool isPositiveSemiDefinite(const Eigen::MatrixXf& in_matrix);
 
 
             /* Function: Fill splines assuming p_cv and p_multi_spec have been filled in the SystStruct*/
@@ -107,17 +107,17 @@ namespace PROfit {
 
 
             /* the fractional covariance that is the sum of all during constructor*/
-            Eigen::MatrixXd fractional_covariance;
+            Eigen::MatrixXf fractional_covariance;
 
             /* names of all systs*/
             std::vector<std::string> spline_names;
-            std::vector<double> spline_lo, spline_hi;
+            std::vector<float> spline_lo, spline_hi;
         private:
             std::unordered_map<std::string, std::pair<size_t, SystType>> syst_map;
             std::vector<Spline> splines;
             [[maybe_unused]] size_t n_splines;
-            std::vector<Eigen::MatrixXd> covmat;
-            std::vector<Eigen::MatrixXd> corrmat;
+            std::vector<Eigen::MatrixXf> covmat;
+            std::vector<Eigen::MatrixXf> corrmat;
             //std::vector<MFA> mfa;
             bool anyspline = false;
             bool anycovar  = true;
