+++ src/PROchi.cxx
@@ -5,13 +5,13 @@ using namespace PROfit;
 
 
 PROchi::PROchi(const std::string tag, const PROconfig *conin, const PROpeller *pin, const PROsyst *systin, const PROsc *oscin, const PROspec &datain, int nparams, int nsyst, EvalStrategy strat, std::vector<float> physics_param_fixed) : PROmetric(), model_tag(tag), config(conin), peller(pin), syst(systin), osc(oscin), data(datain), nparams(nparams), nsyst(nsyst), strat(strat), physics_param_fixed(physics_param_fixed), correlated_systematics(false) {
-    last_value = 0.0; last_param = Eigen::VectorXd::Zero(nparams); 
+    last_value = 0.0; last_param = Eigen::VectorXf::Zero(nparams); 
     fixed_index = -999;
 
     // Build the correlation matrix between priors if configured to
     if (conin->m_mcgen_correlations.size()) {
         correlated_systematics = true;
-        prior_covariance = Eigen::MatrixXd::Identity(nsyst, nsyst);
+        prior_covariance = Eigen::MatrixXf::Identity(nsyst, nsyst);
         for (auto const &t: conin->m_mcgen_correlations) {
           auto itA = std::find(systin->spline_names.begin(), systin->spline_names.end(), std::get<0>(t));
           if (itA == systin->spline_names.end()) {
@@ -37,7 +37,7 @@ PROchi::PROchi(const std::string tag, const PROconfig *conin, const PROpeller *p
     collapsed_stat_covariance = data.Spec().array().matrix().asDiagonal();
 }
 
-float PROchi::Pull(const Eigen::VectorXd &systs) {
+float PROchi::Pull(const Eigen::VectorXf &systs) {
     // No correlations: sum of squares
     if (!correlated_systematics) return systs.array().square().sum();
 
@@ -45,25 +45,25 @@ float PROchi::Pull(const Eigen::VectorXd &systs) {
     return systs.dot(prior_covariance.inverse() * systs);
 }
 
-void PROchi::fixSpline(int fix, double valin){
+void PROchi::fixSpline(int fix, float valin){
     fixed_index=fix;
     fixed_val=valin;
     return;
 }
-double PROchi::operator()(const Eigen::VectorXd &param, Eigen::VectorXd &gradient){
+float PROchi::operator()(const Eigen::VectorXf &param, Eigen::VectorXf &gradient){
     return PROchi::operator()(param, gradient, true);
 }
 
 
-double PROchi::operator()(const Eigen::VectorXd &param, Eigen::VectorXd &gradient, bool rungradient){
+float PROchi::operator()(const Eigen::VectorXf &param, Eigen::VectorXf &gradient, bool rungradient){
     // Get Spectra from FillRecoSpectra
-    Eigen::VectorXd subvector1 = param.segment(0, nparams - nsyst);
+    Eigen::VectorXf subvector1 = param.segment(0, nparams - nsyst);
     std::vector<float> fitparams(subvector1.data(), subvector1.data() + subvector1.size());
     if(fitparams.size() == 0 && physics_param_fixed.size()!=0 ) {
         fitparams = physics_param_fixed;
     }
 
-    Eigen::VectorXd subvector2 = param.segment(nparams - nsyst, nsyst);
+    Eigen::VectorXf subvector2 = param.segment(nparams - nsyst, nsyst);
     std::vector<float> shifts(subvector2.data(), subvector2.data() + subvector2.size());
 
     log<LOG_DEBUG>(L"%1% || Shifts size is %2%") % __func__ % shifts.size();
@@ -73,19 +73,19 @@ double PROchi::operator()(const Eigen::VectorXd &param, Eigen::VectorXd &gradien
     //std::cout<<"Spec "<< result.Spec()<<" .. "<<std::endl;
     //result.plotSpectrum(*config,"TTPT");
 
-    Eigen::MatrixXd inverted_collapsed_full_covariance(config->m_num_bins_total_collapsed,config->m_num_bins_total_collapsed);
+    Eigen::MatrixXf inverted_collapsed_full_covariance(config->m_num_bins_total_collapsed,config->m_num_bins_total_collapsed);
     
     //only calculate a syst covariance if we have any covariance parameters as defined in the xml
     if(syst->GetNCovar()){
 
       // Calculate Full Syst Covariance matrix
-      Eigen::MatrixXd diag = result.Spec().array().matrix().asDiagonal(); 
-      Eigen::MatrixXd full_covariance =  diag*(syst->fractional_covariance)*diag;
+      Eigen::MatrixXf diag = result.Spec().array().matrix().asDiagonal(); 
+      Eigen::MatrixXf full_covariance =  diag*(syst->fractional_covariance)*diag;
       //std::cout<<"Full: "<<full_covariance.size()<<std::endl;
       //std::cout<<full_covariance<<std::endl;
 
       // Collapse Covariance and Spectra 
-      Eigen::MatrixXd collapsed_full_covariance =  CollapseMatrix(*config,full_covariance);  
+      Eigen::MatrixXf collapsed_full_covariance =  CollapseMatrix(*config,full_covariance);  
       log<LOG_DEBUG>(L"%1% || Collapsed second matrix") % __func__;
 
       //std::cout<<"cFull: "<<collapsed_full_covariance.size()<<std::endl;
@@ -105,7 +105,7 @@ double PROchi::operator()(const Eigen::VectorXd &param, Eigen::VectorXd &gradien
       //std::cout<<inverted_collapsed_full_covariance<<std::endl;
 
     // Calculate Chi^2  value
-    Eigen::VectorXd delta  = CollapseMatrix(*config,result.Spec()) - data.Spec(); 
+    Eigen::VectorXf delta  = CollapseMatrix(*config,result.Spec()) - data.Spec(); 
 
     if(!(fixed_index<0)){
         //subvector2[fixed_index]=0;   
@@ -121,31 +121,31 @@ double PROchi::operator()(const Eigen::VectorXd &param, Eigen::VectorXd &gradien
     if(rungradient){
         float dval = 1e-4;
         for (int i = 0; i < nparams; i++) {
-            //Eigen::VectorXd tmpParams = last_param;
-            Eigen::VectorXd tmpParams = param;
+            //Eigen::VectorXf tmpParams = last_param;
+            Eigen::VectorXf tmpParams = param;
             int sgn = ((param(i) - last_param(i)) > 0) - ((param(i) - last_param(i)) < 0);
             if(!sgn) sgn = 1;
             //if(fitparams.size() != 0 && i == 1 && param(i) < -4 + dval) sgn = 1;
             //else if(fitparams.size() != 0 && i == 1 && param(i) > 0 - dval) sgn = -1;
             tmpParams(i) = /*param(i) != last_param(i) ? param(i) :*/ param(i) + sgn * dval;
             
-            Eigen::VectorXd subvector1 = tmpParams.segment(0, nparams - nsyst);
+            Eigen::VectorXf subvector1 = tmpParams.segment(0, nparams - nsyst);
             std::vector<float> fitparams(subvector1.data(), subvector1.data() + subvector1.size());
             if(fitparams.size() == 0 && physics_param_fixed.size() != 0) {
                 fitparams = physics_param_fixed;
             }
-            Eigen::VectorXd subvector2 = tmpParams.segment(nparams - nsyst, nsyst);
+            Eigen::VectorXf subvector2 = tmpParams.segment(nparams - nsyst, nsyst);
             std::vector<float> shifts(subvector2.data(), subvector2.data() + subvector2.size());
             PROspec result = FillRecoSpectra(*config, *peller, *syst, osc, shifts, fitparams, strat != EventByEvent);
             // Calcuate Full Covariance matrix
-            Eigen::MatrixXd inverted_collapsed_full_covariance(config->m_num_bins_total_collapsed,config->m_num_bins_total_collapsed);
+            Eigen::MatrixXf inverted_collapsed_full_covariance(config->m_num_bins_total_collapsed,config->m_num_bins_total_collapsed);
 
             if(syst->GetNCovar()){
 
-                Eigen::MatrixXd diag = result.Spec().array().matrix().asDiagonal(); 
-                Eigen::MatrixXd full_covariance =  diag*(syst->fractional_covariance)*diag;
+                Eigen::MatrixXf diag = result.Spec().array().matrix().asDiagonal(); 
+                Eigen::MatrixXf full_covariance =  diag*(syst->fractional_covariance)*diag;
                 // Collapse Covariance and Spectra 
-                Eigen::MatrixXd collapsed_full_covariance =  CollapseMatrix(*config,full_covariance);  
+                Eigen::MatrixXf collapsed_full_covariance =  CollapseMatrix(*config,full_covariance);  
                 // Invert Collaped Matrix Matrix 
                 inverted_collapsed_full_covariance = (collapsed_full_covariance+collapsed_stat_covariance).inverse();
             }
@@ -154,7 +154,7 @@ double PROchi::operator()(const Eigen::VectorXd &param, Eigen::VectorXd &gradien
                 }
            
             // Calculate Chi^2  value
-            Eigen::VectorXd delta  = CollapseMatrix(*config,result.Spec()) - data.Spec(); 
+            Eigen::VectorXf delta  = CollapseMatrix(*config,result.Spec()) - data.Spec(); 
 
             if(!(fixed_index<0)){
                 //subvector2[fixed_index]=0;
